Added more weapons up to exotic light weapons.
Also fixed EquippableItem to make use of weight instead of return 0.
Also made Item keep track of cost now cost is passed up the constructors
to it instead of subclasses keeping that information.
Also made every weapon and armor require weights and added weight to the
enums.

+++ src/pathfinder/realWorldObject/RealWorldObject.java
@@ -13,7 +13,7 @@ import java.util.List;
 public abstract class RealWorldObject
 {
     private Coordinate coordinate;
-    private long weight;
+    private int weight;
     private final long id;
 
     // perhaps a map or hashset would be better if we want to access these
@@ -31,12 +31,12 @@ public abstract class RealWorldObject
         // idmaker class that has an id counter
     }
 
-    public long getWeight()
+    public int getWeight()
     {
         return weight;
     }
 
-    public void setWeight(long weight)
+    public void setWeight(int weight)
     {
         this.weight = weight;
     }
